Remove more deps from SipHash
