+++ src/IO/UncompressedCache.cpp
@@ -1,4 +1,5 @@
 #include <IO/UncompressedCache.h>
+#include <Common/SipHash.h>
 
 namespace DB
 {
@@ -7,5 +8,14 @@ template class CacheBase<UInt128, UncompressedCacheCell, UInt128TrivialHash, Unc
 UncompressedCache::UncompressedCache(const String & cache_policy, size_t max_size_in_bytes, double size_ratio)
     : Base(cache_policy, max_size_in_bytes, 0, size_ratio)
 {
+}
+
+UInt128 UncompressedCache::hash(const String & path_to_file, size_t offset)
+{
+    SipHash hash;
+    hash.update(path_to_file.data(), path_to_file.size() + 1);
+    hash.update(offset);
+
+    return hash.get128();
 }
 }
