+++ src/Interpreters/HashJoin/KeyGetter.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <Interpreters/HashJoin/HashJoin.h>
+#include <Common/ColumnsHashing.h>
 
 
 namespace DB
